+++ src/react/img-file.ts
